IDA event tests
* fix up readme
* add .bin to gitignore
* fix up dev script
* separation in schema.graphql
* implemented IDAv1 event entity mapping tests
* split cfa event and hol test files, same pattern to be applied when LOC is too great
