+++ packages/subgraph/tests/cfav1/cfav1.helper.ts
@@ -1,25 +1,24 @@
-import { BigInt, Bytes } from "@graphprotocol/graph-ts";
-import { newMockEvent } from "matchstick-as";
+import { Address, BigInt, Bytes } from "@graphprotocol/graph-ts";
+import { assert, newMockEvent } from "matchstick-as";
 import {
     FlowUpdated,
     FlowUpdatedExtension,
     FlowOperatorUpdated,
 } from "../../generated/ConstantFlowAgreementV1/IConstantFlowAgreementV1";
+import { handleFlowUpdated } from "../../src/mappings/cfav1";
+import { getStreamID, ZERO_ADDRESS } from "../../src/utils";
+import { assertEventBaseProperties } from "../assertionHelpers";
+import { DEFAULT_DECIMALS, LIQUIDATION_PERIOD } from "../constants";
 import {
     getAddressEventParam,
     getBigIntEventParam,
     getBytesEventParam,
     getI32EventParam,
+    stringToBytes,
 } from "../converters";
+import { mockedHandleFlowUpdatedRPCCalls } from "../mockedFunctions";
 
-export function getFlowOperatorId(
-    flowOperatorAddress: string,
-    tokenAddress: string,
-    senderAddress: string
-): string {
-    return flowOperatorAddress + "-" + tokenAddress + "-" + senderAddress;
-}
-
+// Mock Event Creators
 export function createFlowUpdatedEvent(
     token: string,
     sender: string,
@@ -74,4 +73,103 @@ export function createFlowOperatorUpdatedEvent(
     newFlowOperatorUpdatedEvent.parameters.push(getBigIntEventParam("flowRateAllowance", flowRateAllowance));
 
     return newFlowOperatorUpdatedEvent;
+}
+
+// Misc Helper Functions
+/**
+ * Create a flowUpdated event and assert the properties were created correctly
+ * @param superToken 
+ * @param tokenName 
+ * @param tokenSymbol 
+ * @param sender 
+ * @param receiver 
+ * @param underlyingToken 
+ * @param expectedType 0 (create), 1 (update) or 2 (delete)
+ * @param expectedOwedDeposit 
+ * @param flowRate 
+ * @param previousSenderFlowRate 
+ * @param previousReceiverFlowRate 
+ * @param isListed 
+ * @returns FlowUpdated event
+ */
+ export function modifyFlowAndAssertFlowUpdatedEventProperties(
+    superToken: string,
+    tokenName: string,
+    tokenSymbol: string,
+    sender: string,
+    receiver: string,
+    underlyingToken: Address,
+    expectedType: string,
+    expectedOwedDeposit: BigInt,
+    flowRate: BigInt,
+    previousSenderFlowRate: BigInt,
+    previousReceiverFlowRate: BigInt,
+    isListed: boolean
+): FlowUpdated {
+    const oldFlowRate = previousSenderFlowRate.abs();
+    const flowRateDelta = flowRate.minus(previousSenderFlowRate);
+    const totalSenderFlowRate = previousSenderFlowRate.minus(flowRateDelta);
+    const totalReceiverFlowRate = previousReceiverFlowRate.plus(flowRateDelta);
+    const userData = stringToBytes("");
+
+    const deposit = getDeposit(flowRate);
+
+    const flowUpdatedEvent = createFlowUpdatedEvent(
+        superToken,
+        sender,
+        receiver,
+        flowRate,
+        totalSenderFlowRate,
+        totalReceiverFlowRate,
+        userData
+    );
+
+    mockedHandleFlowUpdatedRPCCalls(
+        flowUpdatedEvent,
+        superToken,
+        DEFAULT_DECIMALS,
+        tokenName,
+        tokenSymbol,
+        underlyingToken,
+        deposit,
+        expectedOwedDeposit,
+        isListed
+    );
+
+    handleFlowUpdated(flowUpdatedEvent);
+
+    const id = assertEventBaseProperties(
+        flowUpdatedEvent,
+        "FlowUpdated"
+    );
+    const streamId = getStreamID(
+        Address.fromString(sender),
+        Address.fromString(receiver),
+        Address.fromString(superToken),
+        0
+    );
+    assert.fieldEquals("FlowUpdatedEvent", id, "token", superToken);
+    assert.fieldEquals("FlowUpdatedEvent", id, "sender", sender);
+    assert.fieldEquals("FlowUpdatedEvent", id, "receiver", receiver);
+    assert.fieldEquals("FlowUpdatedEvent", id, "flowOperator", ZERO_ADDRESS.toHexString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "flowRate", flowRate.toString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "totalSenderFlowRate", totalSenderFlowRate.toString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "totalReceiverFlowRate", totalReceiverFlowRate.toString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "deposit", deposit.toString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "userData", userData.toHexString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "oldFlowRate", oldFlowRate.toString());
+    assert.fieldEquals("FlowUpdatedEvent", id, "type", expectedType);
+    assert.fieldEquals("FlowUpdatedEvent", id, "stream", streamId);
+
+    return flowUpdatedEvent;
+}
+
+/**
+ * Calculates the deposit amount given a flow rate.
+ * NOTE: We ignore clipping here and only update flow: multiply flow rate by the liquidation period.
+ * @param flowRate
+ * @returns unclipped deposit
+ */
+ export function getDeposit(flowRate: BigInt): BigInt {
+    return flowRate.times(LIQUIDATION_PERIOD);
 }
